test(select): migrate wrapping
