+++ core/src/components/select/test/wrapping/select.e2e.ts-snapshots/select-nowrap-ios-ltr-Mobile-Safari-linux.png
