+++ core/src/components/select/test/wrapping/select.e2e.ts-snapshots/select-nowrap-md-ltr-Mobile-Chrome-linux.png
