+++ core/src/components/select/test/wrapping/select.e2e.ts-snapshots/select-wrap-ios-ltr-Mobile-Firefox-linux.png
