+++ core/src/components/select/test/wrapping/select.e2e.ts-snapshots/select-wrap-md-ltr-Mobile-Chrome-linux.png
